fix: Use sqlx try_from attribute for SQLite INT4 to i64 conversion

+++ server/src/t2/ai.rs
@@ -269,7 +269,7 @@ match_score от 0 до 100 - насколько товар соответств
     struct AIRecommendation {
         product_id: i64,
         price_category: String,
-        match_score: i64,
+        match_score: i32,
         match_reasons: Vec<String>,
     }
