+++ server/src/t2/models.rs
@@ -3,6 +3,7 @@ use sqlx::FromRow;
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Store {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
     pub name: String,
     pub address: String,
@@ -13,7 +14,9 @@ pub struct T2Store {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Employee {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
     pub name: String,
     pub code: String,
@@ -34,6 +37,7 @@ pub struct T2EmployeeWithStores {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Category {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
     pub name: String,
     pub icon: Option<String>,
@@ -42,24 +46,31 @@ pub struct T2Category {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Tag {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
     pub name: String,
     pub color: String,
     pub description: Option<String>,
+    #[sqlx(try_from = "i32")]
     pub priority: i64,
     pub created_at: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Product {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
+    #[sqlx(try_from = "i32")]
     pub category_id: i64,
     pub name: String,
     pub brand: Option<String>,
     pub model: Option<String>,
     pub price: f64,
+    #[sqlx(try_from = "i32")]
     pub quantity: i64,
     pub image_url: Option<String>,
     pub created_at: String,
@@ -86,7 +97,9 @@ pub struct T2ProductWithDetails {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2ProductSpec {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub product_id: i64,
     pub spec_name: String,
     pub spec_value: String,
@@ -94,13 +107,15 @@ pub struct T2ProductSpec {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Tariff {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
     pub name: String,
     pub price: f64,
-    pub minutes: Option<i64>,
-    pub sms: Option<i64>,
-    pub gb: Option<i64>,
+    pub minutes: Option<i32>,
+    pub sms: Option<i32>,
+    pub gb: Option<i32>,
     pub unlimited_t2: bool,
     pub unlimited_internet: bool,
     pub unlimited_sms: bool,
@@ -113,7 +128,9 @@ pub struct T2Tariff {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Service {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
     pub name: String,
     pub price: f64,
@@ -124,8 +141,11 @@ pub struct T2Service {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Sale {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub store_id: i64,
+    #[sqlx(try_from = "i32")]
     pub employee_id: i64,
     pub customer_request: Option<String>,
     pub customer_audio_url: Option<String>,
@@ -151,20 +171,26 @@ pub struct T2SaleWithDetails {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2SaleItem {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub sale_id: i64,
     pub item_type: String,
+    #[sqlx(try_from = "i32")]
     pub item_id: i64,
     pub item_name: String,
     pub item_details: Option<String>,
     pub price: f64,
+    #[sqlx(try_from = "i32")]
     pub quantity: i64,
     pub created_at: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Session {
+    #[sqlx(try_from = "i32")]
     pub id: i64,
+    #[sqlx(try_from = "i32")]
     pub employee_id: i64,
     pub token: String,
     pub expires_at: String,
@@ -203,7 +229,7 @@ pub struct CreateProductRequest {
     pub brand: Option<String>,
     pub model: Option<String>,
     pub price: f64,
-    pub quantity: Option<i64>,
+    pub quantity: Option<i32>,
     pub image_url: Option<String>,
     pub specs: Vec<ProductSpecInput>,
     pub tag_ids: Vec<i64>,
@@ -221,7 +247,7 @@ pub struct UpdateProductRequest {
     pub brand: Option<String>,
     pub model: Option<String>,
     pub price: Option<f64>,
-    pub quantity: Option<i64>,
+    pub quantity: Option<i32>,
     pub image_url: Option<String>,
     pub specs: Option<Vec<ProductSpecInput>>,
     pub tag_ids: Option<Vec<i64>>,
@@ -232,7 +258,7 @@ pub struct CreateTagRequest {
     pub name: String,
     pub color: String,
     pub description: Option<String>,
-    pub priority: Option<i64>,
+    pub priority: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -240,16 +266,16 @@ pub struct UpdateTagRequest {
     pub name: Option<String>,
     pub color: Option<String>,
     pub description: Option<String>,
-    pub priority: Option<i64>,
+    pub priority: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct CreateTariffRequest {
     pub name: String,
     pub price: f64,
-    pub minutes: Option<i64>,
-    pub sms: Option<i64>,
-    pub gb: Option<i64>,
+    pub minutes: Option<i32>,
+    pub sms: Option<i32>,
+    pub gb: Option<i32>,
     pub unlimited_t2: Option<bool>,
     pub unlimited_internet: Option<bool>,
     pub unlimited_sms: Option<bool>,
@@ -262,9 +288,9 @@ pub struct CreateTariffRequest {
 pub struct UpdateTariffRequest {
     pub name: Option<String>,
     pub price: Option<f64>,
-    pub minutes: Option<i64>,
-    pub sms: Option<i64>,
-    pub gb: Option<i64>,
+    pub minutes: Option<i32>,
+    pub sms: Option<i32>,
+    pub gb: Option<i32>,
     pub unlimited_t2: Option<bool>,
     pub unlimited_internet: Option<bool>,
     pub unlimited_sms: Option<bool>,
@@ -299,7 +325,7 @@ pub struct CustomerRequest {
 pub struct ProductRecommendation {
     pub product: T2ProductWithDetails,
     pub price_category: String,
-    pub match_score: i64,
+    pub match_score: i32,
     pub match_reasons: Vec<String>,
 }
 
@@ -351,7 +377,7 @@ pub struct CreateSaleRequest {
 pub struct SaleItemInput {
     pub item_type: String,
     pub item_id: i64,
-    pub quantity: Option<i64>,
+    pub quantity: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
